manage doctor is done

+++ src/Pages/Dashboard/DoctorRow.js
@@ -5,19 +5,22 @@ const DoctorRow = ({ docotor, index, refetch }) => {
   const { _id,name, specility, image } = docotor;
   const deleteDoctor = _id => {
       console.log(_id);
-      fetch(`http://localhost:5000/doctors/${_id}`,{
-          method:'DELETE',
-          headers: {
-            authorization: `Bearer ${localStorage.getItem("accessToken")}`,
-          }
-      }).then((res) => res.json())
-      .then((result) => {
-          console.log(result);
-          if(result.deletedCount){
-              toast.success(`doctor:${name} is deleteded`);
-              refetch();
-          }
-      })
+      const isDelete = window.confirm(`Are you sure You wan't to delete this doctor?`)
+      if(isDelete){
+        fetch(`http://localhost:5000/doctors/${_id}`,{
+            method:'DELETE',
+            headers: {
+              authorization: `Bearer ${localStorage.getItem("accessToken")}`,
+            }
+        }).then((res) => res.json())
+        .then((result) => {
+            console.log(result);
+            if(result.deletedCount){
+                toast.success(`doctor:${name} is deleteded`);
+                refetch();
+            }
+        })
+      }
   }
   return (
     <tr>
@@ -31,7 +34,7 @@ const DoctorRow = ({ docotor, index, refetch }) => {
       </td>
       <td>{name}</td>
       <td>{specility}</td>
-      <td> <button className="btn btn-xs btn-error" onClick={() => deleteDoctor(_id)}>Delete</button> </td>
+      <td> <label htmlFor="delete-modal"  className="btn  modal-button btn-xs btn-error" onClick={() => deleteDoctor(_id)}>Delete</label> </td>
     </tr>
   );
 };
